+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGE LOG: User sessions (z4m_usersessions)
 
+## Version 1.4, 2025-08-11
+- CHANGE: anonymous sessions (sessions created when user download the login page) have now a lifetime of 5 minutes (new `MOD_Z4M_USERSESSIONS_ANONYMOUS_SESSION_LIFETIME` PHP constant) and are cleaned when have expired.
+- CHANGE: the value of the new `MOD_Z4M_USERSESSIONS_ANONYMOUS_SESSION_LIFETIME` PHP constant is displayed in the session configuration modal dialog.
+
 ## Version 1.3, 2025-06-27
 - BUG FIXING: E_WARNING - unserialize(): Extra data starting at offset 406 of 1042 bytes - ./engine/modules/z4m_usersessions/mod/UserSessionFile.php(78)